+++ app/src/main/java/com/swiftcause/swiftcause_android/MainActivity.kt
@@ -4,6 +4,7 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
+import com.swiftcause.swiftcause_android.ui.navigation.NavController
 
 
 class MainActivity : ComponentActivity() {
