--- app/src/main/java/com/swiftcause/swiftcause_android/Routes.kt
@@ -1,6 +1,0 @@
-package com.swiftcause.swiftcause_android
-
-object Routes {
-    val welcomeScreen = "welcome_screen"
-    val homeScreen = "home_screen"
-}
