+++ app/src/main/java/com/swiftcause/swiftcause_android/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.swiftcause.swiftcause_android.ui.theme
+package com.swiftcause.swiftcause_android.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
