+++ app/src/main/java/com/swiftcause/swiftcause_android/ui/navigation/Routes.kt
@@ -1,0 +1,6 @@
+package com.swiftcause.swiftcause_android.ui.navigation
+
+object Routes {
+    val campaignListScreen = "campaign_list_screen"
+    val loginScreen = "login_screen"
+}
