+++ app/src/main/java/com/swiftcause/swiftcause_android/ui/screen/welcome/LoginScreen.kt
@@ -1,10 +1,11 @@
-package com.swiftcause.swiftcause_android.screens
+package com.swiftcause.swiftcause_android.ui.screen.welcome
 
 import android.app.Activity
 import android.util.Log
 import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.result.contract.ActivityResultContracts
-import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.Button
 import androidx.compose.material3.Text
@@ -16,26 +17,34 @@ import androidx.navigation.NavController
 import com.firebase.ui.auth.AuthUI
 import com.firebase.ui.auth.IdpResponse
 import com.google.firebase.auth.FirebaseAuth
-import com.swiftcause.swiftcause_android.Routes
+import com.swiftcause.swiftcause_android.ui.navigation.Routes
 
 @Composable
-fun WelcomeScreen(navController: NavController){
+fun LoginScreen(navController: NavController) {
     val context = LocalContext.current;
     val launcher = rememberLauncherForActivityResult(
         contract = ActivityResultContracts.StartActivityForResult()
     ) { result ->
         val response = IdpResponse.fromResultIntent(result.data)
-        if (result.resultCode == Activity.RESULT_OK){
+        if (result.resultCode == Activity.RESULT_OK) {
             // login successful
             val user = FirebaseAuth.getInstance().currentUser
-            navController.navigate(Routes.homeScreen + "/${user?.displayName}")
-            Log.d("Fire-Auth","the user is , ${user?.email}, ${user?.uid}")
-        }else{
-            Log.d("Fire-Auth", "the login was not successful due to : ${response?.error?.errorCode}")
+            navController.navigate(Routes.campaignListScreen + "/${user?.displayName}")
+            Log.d("Fire-Auth", "the user is , ${user?.email}, ${user?.uid}")
+        } else {
+            Log.d(
+                "Fire-Auth",
+                "the login was not successful due to : ${response?.error?.errorCode}"
+            )
         }
     }
 
-    Box(modifier = Modifier.fillMaxSize(), contentAlignment = Alignment.Center){
+    Column(
+        modifier = Modifier.fillMaxSize(),
+        verticalArrangement = Arrangement.SpaceAround,
+        horizontalAlignment = Alignment.CenterHorizontally
+    ) {
+        Text("Welcome to SwiftCause (welcome screen)")
         Button(onClick = {
             val provider = arrayListOf(
                 AuthUI.IdpConfig.GoogleBuilder().build(),
@@ -49,7 +58,7 @@ fun WelcomeScreen(navController: NavController){
                 .build()
 
             launcher.launch(signInIntent)
-        }){
+        }) {
             Text("Sign in")
         }
     }
